+++ playground/App.vue
@@ -19,6 +19,11 @@
             </div>
             <InputExampleOne :tabindex="4"></InputExampleOne>
 
+            <div class="row mb-2 mt-5">
+                <h3>Vue Textarea Control</h3>
+                <p>Textarea control that can capitalize, and restrict keys.</p>
+            </div>
+            <TextareaExampleOne :tabindex="4"></TextareaExampleOne>
         </div>
     </div>
 </template>
@@ -27,5 +32,6 @@ import ExampleOne from "./ExampleOne.vue"
 import ExampleTwo from "./ExampleTwo.vue"
 import ExampleThree from "./ExampleThree.vue"
 import InputExampleOne from "./InputExampleOne.vue";
+import TextareaExampleOne from "./TextareaExampleOne.vue";
 </script>
 <style scoped></style>