+++ src/index.ts
@@ -1,6 +1,7 @@
 import type { App } from "vue"
 import VueSelect from "./VueSelect.vue"
 import VueInput from "./VueInput.vue"
+import VueTextarea from "./VueTextarea.vue"
 export type { FormatFunction, FilterFunction } from "./types"
 
 
@@ -8,7 +9,8 @@ export default {
     install: (app: App) => {
         app.component('VueSelect', VueSelect) 
         app.component('VueInput', VueInput)
+        app.component('VueTextarea', VueTextarea)
     }
 }
 
-export { VueSelect, VueInput }
+export { VueSelect, VueInput, VueTextarea }